update: rename components
