+++ src/components/App.js
@@ -1,5 +1,5 @@
 import React, { PureComponent } from 'react';
-import Restaurants from './restaurants';
+import Restaurants from './Restaurants/Restaurants';
 
 export default class App extends PureComponent {
   render() {
