+++ src/components/Restaurants/Product.js
@@ -1,17 +1,19 @@
 import React from 'react';
-import counter from '../hocs/counter';
+import counter from '../../hocs/Counter';
 
-import minus from '../icons/minus.svg';
-import Plus from '../icons/plus';
+import minus from '../../icons/minus.svg';
+import Plus from '../../icons/plus';
 
 function Product(props) {
   const { amount, decrement, increment } = props;
-
+  const style = {
+    cursor: amount === 0 ? 'not-allowed' : 'pointer',
+  };
   return (
     <div>
       <p>{props.product.name}</p>
       <p>{props.product.price} $</p>
-      <button onClick={decrement}>
+      <button onClick={decrement} disabled={amount === 0} style={style}>
         <img src={minus} alt="minus" />
       </button>
       {amount}
