+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 
-import App from './components/app';
+import App from './components/App';
 
 import { restaurants } from './fixtures';
